added routes for landing pages, questions, results and 404
