+++ app/package.json
@@ -31,6 +31,7 @@
     "prettier": "^2.8.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.2",
     "react-scripts": "^5.0.1",
     "serve": "^14.1.2",
     "web-vitals": "^2.1.4"
