+++ app/src/pages/LandingPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const LandingPage: React.FC = () => {
+  return <p>Page Results - LandingPage</p>
+}
+
+export default LandingPage
