+++ app/src/pages/NotFound.css
@@ -1,0 +1,7 @@
+.header-404 {
+    font-family:Georgia, 'Times New Roman', Times, serif;
+    font-size: 36px;
+    font-weight: 500;
+    letter-spacing: -2px;
+    line-height: 45px;
+}
