+++ app/src/pages/NotFound.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import './NotFound.css'
+
+const NotFound: React.FC = () => {
+  return <h2 className='header-404'>Page not found</h2>
+}
+
+export default NotFound
