+++ app/src/pages/Result.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Result: React.FC = () => {
+  return <p>Page Results - Placeholder</p>
+}
+
+export default Result
